backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
flask-cors==4.0.0
pytest==7.4.2
